ArmIOSim and changed ArmIO degrees to double(was int)

+++ src/main/java/frc/robot/subsystems/Arm/ArmIO.java
@@ -5,10 +5,10 @@ import org.littletonrobotics.junction.AutoLog;
 public interface ArmIO {
     @AutoLog
     public static class ArmIOInputs {
-    public int Degrees;
+    public double Degrees;
         
     }
-    public default void setDegrees(int Degrees) {}
+    public default void setDegrees(Double Degrees) {}
 
     public default void updateInputs(ArmIOInputs inputs) {}
